+++ js/app/controllers/sidebar/sidebarcontroller.js
@@ -13,7 +13,7 @@ angular.module('Music').controller('SidebarController', [
 	'$rootScope', '$scope', '$timeout',
 	function ($rootScope, $scope, $timeout) {
 
-		$scope.follow = (localStorage.getItem('oc_music_details_follow_playback') === 'true');
+		$scope.follow = (OCA.Music.Storage.get('details_follow_playback') === 'true');
 
 		$scope.contentType = null;
 		$scope.contentId = null;
@@ -106,7 +106,7 @@ angular.module('Music').controller('SidebarController', [
 
 		$scope.toggleFollow = function() {
 			$scope.follow = !$scope.follow;
-			localStorage.setItem('oc_music_details_follow_playback', $scope.follow.toString());
+			OCA.Music.Storage.set('details_follow_playback', $scope.follow.toString());
 
 			// If "follow playback" was enabled and the currently shown track doesn't match currently
 			// playing track, then immediately switch to the details of the playing track.